pyproject.toml
[tool.poetry]
name = "salary-service"
version = "0.1.0"
description = "REST API сервис для просмотра зарплаты сотрудников"
authors = ["sapronov.dmitrii19@yandex.com"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.9.1"
alembic = "^1.16.2"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
